Move shim service into top lvl package
